# internal/interfaces/http/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/accounts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lista todas as contas do usuário autenticado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "List accounts",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Número máximo de resultados (default: 100, max: 200)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Número de resultados para pular (default: 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de contas",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_vasconcellos_financial-control_internal_domain_dto.AccountResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cria uma nova conta bancária para o usuário autenticado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Create a new account",
                "parameters": [
                    {
                        "description": "Dados da conta",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_vasconcellos_financial-control_internal_domain_dto.CreateAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Conta criada com sucesso",
                        "schema": {
                            "$ref": "#/definitions/github_com_vasconcellos_financial-control_internal_domain_dto.AccountResponse"
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/accounts/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove uma conta existente",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Delete an account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da conta",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Conta removida com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Conta não encontrada",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atualiza uma conta existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Update an account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da conta",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados atualizados",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_vasconcellos_financial-control_internal_domain_dto.UpdateAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conta atualizada",
                        "schema": {
                            "$ref": "#/definitions/github_com_vasconcellos_financial-control_internal_domain_dto.AccountResponse"
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Conta não encontrada",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Autentica um usuário e retorna tokens de acesso",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Authenticate user",
                "parameters": [
                    {
                        "description": "Credenciais de login",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_vasconcellos_financial-control_internal_domain_dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tokens de autenticação",
                        "schema": {
                            "$ref": "#/definitions/github_com_vasconcellos_financial-control_internal_domain_dto.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_vasconcellos_financial-control_internal_domain_dto.AccountResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "currency": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "github_com_vasconcellos_financial-control_internal_domain_dto.CreateAccountRequest": {
            "type": "object",
            "required": [
                "currency",
                "name",
                "type"
            ],
            "properties": {
                "balance": {
                    "type": "number"
                },
                "currency": {
                    "type": "string",
                    "enum": [
                        "USD",
                        "EUR",
                        "CHF",
                        "GBP",
                        "BRL"
                    ]
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "checking",
                        "savings",
                        "credit",
                        "cash"
                    ]
                }
            }
        },
...
        "github_com_vasconcellos_financial-control_internal_domain_dto.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "minLength": 8
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_vasconcellos_financial-control_internal_domain_dto.LoginResponse": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "expiresIn": {
                    "type": "integer"
                },
                "idToken": {
                    "type": "string"
                },
                "refreshToken": {
                    "type": "string"
                },
                "tokenType": {
                    "type": "string"
                }
            }
        },
...
        "github_com_vasconcellos_financial-control_internal_domain_dto.UpdateAccountRequest": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string",
                    "enum": [
                        "USD",
                        "EUR",
                        "CHF",
                        "GBP",
                        "BRL"
                    ]
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "internal_interfaces_http_handler.ErrorResponse": {
            "description": "Standard error response structure",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
    }
}